refactor(config): move TestResolveDogstatsdAddr back to tracer package
Restore TestResolveDogstatsdAddr in ddtrace/tracer/option_test.go and remove the
duplicated test from internal/config/config_helpers_test.go. Tests in the tracer
package that exercise config functionality should remain there; the test now
calls through the public *internalconfig.Config API instead of the package-local
helper.
